refactor: Dropdown 커스텀훅 추출
PortDropdown, TerminalDropdown 중복 해결

LR-24 #time 20m
